refactor slice np mcmc, seed sir, add tests.
