+++ sbi/inference/posteriors/mcmc_posterior.py
@@ -1,11 +1,9 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 from functools import partial
-from math import ceil
 from typing import Any, Callable, Dict, Optional, Union
 from warnings import warn
 
-import numpy as np
 import torch
 import torch.distributions.transforms as torch_tf
 from joblib import Parallel, delayed
@@ -15,15 +13,13 @@ from torch import Tensor
 from torch import multiprocessing as mp
 from tqdm import tqdm
 
-from sbi import utils as utils
 from sbi.inference.posteriors.base_posterior import NeuralPosterior
 from sbi.samplers.mcmc import (
     IterateParameters,
     Slice,
-    SliceSampler,
-    SliceSamplerVectorized,
     proposal_init,
     sir,
+    slice_np_parallized,
 )
 from sbi.simulators.simutils import tqdm_joblib
 from sbi.types import Shape, TorchTransform
@@ -75,8 +71,7 @@ class MCMCPosterior(NeuralPosterior):
                 with weights proportional to the `potential_fn`-value.
             init_strategy_num_candidates: Number of candidates to to find init
                 locations in `init_strategy=sir`.
-            num_workers: number of cpu core used to parallelize initialization of mcmc
-                parameters.
+            num_workers: number of cpu cores used to parallelize mcmc
             device: Training device, e.g., "cpu", "cuda" or "cuda:0". If None,
                 `potential_fn.device` is used.
             x_shape: Shape of a single simulator output. If passed, it is used to check
@@ -237,28 +232,9 @@ class MCMCPosterior(NeuralPosterior):
         )
         self.potential_ = self._prepare_potential(method)  # type: ignore
 
-        init_fn = self._build_mcmc_init_fn(
-            self.proposal,
-            self.potential_fn,
-            transform=self.theta_transform,
-            init_strategy=init_strategy,  # type: ignore
+        initial_params = self._get_initial_params(
+            init_strategy, num_chains, num_workers, show_progress_bars  # type: ignore
         )
-
-        # Generate initial params parallelized over num_workers.
-        with tqdm_joblib(
-            tqdm(
-                range(num_chains),  # type: ignore
-                disable=not show_progress_bars,
-                desc=f"Generating {num_chains} MCMC inits with {num_workers} workers.",
-                total=num_chains,
-            )
-        ) as _:
-            initial_params = torch.cat(
-                Parallel(n_jobs=num_workers)(
-                    delayed(init_fn)() for _ in range(num_chains)  # type: ignore
-                )
-            )
-
         num_samples = torch.Size(sample_shape).numel()
 
         track_gradients = method in ("hmc", "nuts")
@@ -328,6 +304,66 @@ class MCMCPosterior(NeuralPosterior):
         else:
             raise NotImplementedError
 
+    def _get_initial_params(
+        self,
+        init_strategy: str,
+        num_chains: int,
+        num_workers: int,
+        show_progress_bars: bool,
+    ) -> Tensor:
+        """Return initial parameters for MCMC obtained with given init strategy.
+
+        Parallelizes across CPU cores only for SIR.
+
+        Args:
+            init_strategy: Specifies the initialization method. Either of
+                [`proposal`|`sir`|`latest_sample`].
+            num_chains: number of MCMC chains, generates initial params for each
+            num_workers: number of CPU cores for parallization
+            show_progress_bars: whether to show progress bars for SIR init
+
+        Returns:
+            Tensor: initial parameters, one for each chain
+        """
+        # Build init function
+        init_fn = self._build_mcmc_init_fn(
+            self.proposal,
+            self.potential_fn,
+            transform=self.theta_transform,
+            init_strategy=init_strategy,  # type: ignore
+        )
+
+        # Parallelize inits for SIR only.
+        if num_workers > 1 and init_strategy == "sir":
+
+            def seeded_init_fn(seed):
+                torch.manual_seed(seed)
+                return init_fn()
+
+            seeds = torch.randint(high=2**31, size=(num_workers,))
+
+            # Generate initial params parallelized over num_workers.
+            with tqdm_joblib(
+                tqdm(
+                    range(num_chains),  # type: ignore
+                    disable=not show_progress_bars,
+                    desc=f"""Generating {num_chains} MCMC inits with {num_workers} 
+                         workers.""",
+                    total=num_chains,
+                )
+            ):
+                initial_params = torch.cat(
+                    Parallel(n_jobs=num_workers)(
+                        delayed(seeded_init_fn)(seed) for seed in seeds
+                    )
+                )
+        else:
+            initial_params = torch.cat(
+                [init_fn() for _ in range(num_chains)]  # type: ignore
+            )
+
+        return initial_params
+
     def _slice_np_mcmc(
         self,
         num_samples: int,
@@ -349,64 +385,26 @@ class MCMCPosterior(NeuralPosterior):
             warmup_steps: Initial number of samples to discard.
             vectorized: Whether to use a vectorized implementation of
                 the Slice sampler (still experimental).
+            num_workers: number of CPU cores to use
+            seed: seed that will be used to generate sub-seeds for each worker
             show_progress_bars: Whether to show a progressbar during sampling;
                 can only be turned off for vectorized sampler.
 
         Returns: Tensor of shape (num_samples, shape_of_single_theta).
         """
-        num_chains = initial_params.shape[0]
-        dim_samples = initial_params.shape[1]
 
-        if not vectorized:  # Sample all chains sequentially
-            all_samples = []
+        num_chains, dim_samples = initial_params.shape
 
-            # Define run function for given input.
-            def run_slice_np(inits):
-                posterior_sampler = SliceSampler(
-                    utils.tensor2numpy(inits).reshape(-1),
-                    lp_f=potential_function,
-                    thin=thin,
-                    verbose=show_progress_bars,
-                )
-                if warmup_steps > 0:
-                    posterior_sampler.gen(int(warmup_steps))
-                return posterior_sampler.gen(ceil(num_samples / num_chains))
-
-            # Run parallelized.
-            all_samples = Parallel(n_jobs=num_workers)(
-                delayed(run_slice_np)(initial_params[c, :]) for c in range(num_chains)
-            )
-
-            all_samples = np.stack(all_samples).astype(np.float32)
-            samples = torch.from_numpy(all_samples)  # chains x samples x dim
-        else:  # Sample all chains at the same time
-
-            # Define local function to run a batch of chains vectorized.
-            def run_slice_np_vectorized(initial_params_batch):
-                posterior_sampler = SliceSamplerVectorized(
-                    init_params=utils.tensor2numpy(initial_params_batch),
-                    log_prob_fn=potential_function,
-                    num_chains=initial_params_batch.shape[0],
-                    verbose=show_progress_bars,
-                )
-                warmup_ = warmup_steps * thin
-                num_samples_ = ceil((num_samples * thin) / num_chains)
-                samples = posterior_sampler.run(warmup_ + num_samples_)
-                samples = samples[:, warmup_:, :]  # discard warmup steps
-                samples = samples[:, ::thin, :]  # thin chains
-                samples = torch.from_numpy(samples)  # chains x samples x dim
-                return samples
-
-            # Parallize over batch of chains.
-            initial_params_in_batches = torch.split(
-                initial_params, ceil(num_chains / num_workers), dim=0
-            )
-            all_samples = Parallel(n_jobs=num_workers)(
-                delayed(run_slice_np_vectorized)(initial_params_batch)
-                for initial_params_batch in initial_params_in_batches
-            )
-            all_samples = np.stack(all_samples).astype(np.float32)
-            samples = torch.from_numpy(all_samples)  # chains x samples x dim
+        samples = slice_np_parallized(
+            potential_function,
+            initial_params,
+            num_samples,
+            thin=thin,
+            warmup_steps=warmup_steps,
+            vectorized=vectorized,
+            num_workers=num_workers,
+            show_progress_bars=show_progress_bars,
+        )
 
         # Save sample as potential next init (if init_strategy == 'latest_sample').
         self._mcmc_init_params = samples[:, -1, :].reshape(num_chains, dim_samples)
