+++ sbi/samplers/__init__.py
@@ -1,3 +1,2 @@
 from sbi.samplers.mcmc.slice import Slice
-from sbi.samplers.mcmc.slice_numpy import SliceSampler
-from sbi.samplers.mcmc.slice_numpy_vectorized import SliceSamplerVectorized
+from sbi.samplers.mcmc.slice_numpy import SliceSampler, SliceSamplerVectorized
