+++ sbi/samplers/mcmc/__init__.py
@@ -1,4 +1,7 @@
 from sbi.samplers.mcmc.init_strategy import IterateParameters, proposal_init, sir
 from sbi.samplers.mcmc.slice import Slice
-from sbi.samplers.mcmc.slice_numpy import SliceSampler
-from sbi.samplers.mcmc.slice_numpy_vectorized import SliceSamplerVectorized
+from sbi.samplers.mcmc.slice_numpy import (
+    SliceSampler,
+    SliceSamplerVectorized,
+    slice_np_parallized,
+)
